move exception to separate java file
