+++ .gitignore
@@ -1,3 +1,4 @@
 .vagrant/
 ubuntu-xenial-16.04-cloudimg-console.log
 *.o
+*.swp
